backend/requirements.txt
alembic==1.14.0
flask==3.1.0
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1